separate QuantityButton
