+++ src/components/QuantityButton/QuantityButton.module.css
@@ -1,0 +1,56 @@
+.cartDetailQuantity {
+  width: min-content;
+}
+
+.quantity_button {
+  padding: 6px 14px;
+  background-color: #363f4e;
+  display: flex;
+  border:none;
+}
+
+.quantity_button:hover {
+  background-color: #ea3c31;
+}
+
+.quantity_button svg {
+  color: white!important;
+}
+
+.quantity_in_cart {
+  padding: 0;
+  width: auto;
+  border: none!important;
+}
+
+.quantity_in_cart input {
+  width: 45px;
+  border-radius: 0!important;
+  border: 1px solid var(--bs-border-color);
+  text-align: center;
+}
+
+.quantity_in_cart input:focus-visible {
+  outline: 0;
+}
+
+@media screen and (max-width:992px) {
+  .cartDetailQuantity {
+    display: flex;
+    align-items: flex-end!important;
+    flex-direction: column;
+  }
+  .quantity_in_cart {
+    width: auto;
+  }
+  .quantity_button {
+    padding: 7px 9px;
+  }
+}
+
+@media screen and (max-width:576px) {
+  .quantity_button {
+    padding: 7px 9px;
+  }
+}
+
